requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
facenet-pytorch==2.6.0
filelock==3.18.0
fonttools==4.58.4
fsspec==2025.5.1
idna==3.10
imagecorruptions-imaug==1.1.3
imageio==2.37.0
imaug==0.4.1
Jinja2==3.1.6
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pillow==10.2.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
requests==2.32.4
scikit-image==0.25.2
scipy==1.15.3
shapely==2.1.1
six==1.17.0
sympy==1.14.0
tifffile==2025.5.10
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
triton==2.2.0
typing_extensions==4.14.0
urllib3==2.5.0
